[config] Move genesis config to config crate.
