+++ benchmarks/src/chain.rs
@@ -6,14 +6,14 @@ use parking_lot::RwLock;
 use rand::prelude::*;
 use starcoin_account_api::AccountInfo;
 use starcoin_chain::BlockChain;
-use starcoin_config::{temp_path, DataDirPath, RocksdbConfig};
+use starcoin_config::{temp_path, ChainNetwork, DataDirPath, RocksdbConfig};
 use starcoin_consensus::Consensus;
 use starcoin_genesis::Genesis;
 use starcoin_storage::cache_storage::CacheStorage;
 use starcoin_storage::db_storage::DBStorage;
 use starcoin_storage::storage::StorageInstance;
 use starcoin_storage::Storage;
-use starcoin_vm_types::genesis_config::{ChainNetwork, ConsensusStrategy};
+use starcoin_vm_types::genesis_config::ConsensusStrategy;
 use std::sync::Arc;
 use traits::{ChainReader, ChainWriter};
