+++ benchmarks/src/lib.rs
@@ -4,8 +4,8 @@
 use starcoin_executor::{peer_to_peer_txn_sent_as_association, DEFAULT_EXPIRATION_TIME};
 
 use crypto::ed25519::random_public_key;
+use starcoin_config::ChainNetwork;
 use starcoin_vm_types::account_address;
-use starcoin_vm_types::genesis_config::ChainNetwork;
 use starcoin_vm_types::transaction::authenticator::AuthenticationKey;
 use types::transaction::SignedUserTransaction;
