+++ cmd/miner_client/src/cpu_solver.rs
@@ -6,9 +6,9 @@ use futures_channel::mpsc;
 use futures_channel::mpsc::{unbounded, UnboundedSender};
 use logger::prelude::*;
 use rand::Rng;
-use starcoin_config::{ConsensusStrategy, MinerClientConfig, TimeService};
-use starcoin_types::block::BlockHeaderExtra;
+use starcoin_config::{MinerClientConfig, TimeService};
 use starcoin_types::U256;
+use starcoin_types::{block::BlockHeaderExtra, genesis_config::ConsensusStrategy};
 use std::ops::Range;
 use std::sync::Arc;
 use std::thread;
