+++ cmd/replay/src/main.rs
@@ -6,12 +6,12 @@ use starcoin_chain::verifier::Verifier;
 use starcoin_chain::verifier::{BasicVerifier, ConsensusVerifier, FullVerifier, NoneVerifier};
 use starcoin_chain::BlockChain;
 use starcoin_config::RocksdbConfig;
+use starcoin_config::{BuiltinNetworkID, ChainNetwork};
 use starcoin_genesis::Genesis;
 use starcoin_storage::cache_storage::CacheStorage;
 use starcoin_storage::db_storage::DBStorage;
 use starcoin_storage::storage::StorageInstance;
 use starcoin_storage::Storage;
-use starcoin_vm_types::genesis_config::{BuiltinNetworkID, ChainNetwork};
 use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::SystemTime;
