+++ miner/src/create_block_template/test_create_block_template.rs
@@ -10,12 +10,12 @@ use consensus::Consensus;
 use logger::prelude::*;
 use starcoin_account_api::AccountInfo;
 use starcoin_account_service::AccountService;
+use starcoin_config::ChainNetworkID;
 use starcoin_config::{temp_path, NodeConfig, StarcoinOpt};
 use starcoin_genesis::Genesis as StarcoinGenesis;
 use starcoin_service_registry::{RegistryAsyncService, RegistryService};
 use starcoin_storage::BlockStore;
 use starcoin_txpool::TxPoolService;
-use starcoin_vm_types::genesis_config::ChainNetworkID;
 use starcoin_vm_types::time::MockTimeService;
 use std::sync::Arc;
 use traits::{ChainReader, ChainWriter};
