+++ testsuite/tests/steps/node.rs
@@ -2,9 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 use crate::MyWorld;
 use cucumber::{Steps, StepsBuilder};
-use starcoin_config::{NodeConfig, StarcoinOpt};
+use starcoin_config::{ChainNetworkID, NodeConfig, StarcoinOpt};
 use starcoin_logger::prelude::*;
-use starcoin_vm_types::genesis_config::ChainNetworkID;
 use std::path::PathBuf;
 use std::sync::Arc;
