+++ vm/types/src/on_chain_config/mod.rs
@@ -24,7 +24,7 @@ mod vm_config;
 pub use self::{
     consensus_config::{consensus_config_type_tag, ConsensusConfig, CONSENSUS_CONFIG_IDENTIFIER},
     dao_config::DaoConfig,
-    genesis_gas_schedule::{INITIAL_GAS_SCHEDULE, TEST_GAS_SCHEDULE},
+    genesis_gas_schedule::init_cost_table,
     version::{version_config_type_tag, Version, VERSION_CONFIG_IDENTIFIER},
     vm_config::{vm_config_type_tag, VMConfig, VMPublishingOption, SCRIPT_HASH_LENGTH},
 };
